+++ udata_gouvfr/migrations/2018-04-18-migrate-active-users.js
@@ -1,0 +1,10 @@
+/*
+ * Ensure active users have a confirmed_at date set.
+ */
+
+const result = db.user.update(
+    {active: true, confirmed_at: null},
+    {$set: {confirmed_at: new Date()}},
+    {multi: true}
+);
+print(`${result.nModified} users migrated.`);